old-tests/submission/W3C/domhtml/HTMLBRElement01.html

<!DOCTYPE html>
<html><head>
<meta charset="utf-8">
<title>NIST DOM HTML Test - BR</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../common/domtests-common.js"></script>
<link href="http://www.w3.org/TR/html5/obsolete.html#dom-br-clear" rel="help">
<link rel="author" title="NIST">
</head>
<body>
<p>
<br clear="none">
</p>


<script>
var domtest = async_test("The clear IDL attribute of the br element must reflect the content attribute of the same name.");

on_event(window, "load", domtest.step_func(function (dom_event) {
    var nodeList;
      var testNode;
      var vclear;
      var doc;
      doc = window.document;
      nodeList = doc.getElementsByTagName("br");
      assert_equals(nodeList.length,1);
testNode = nodeList.item(0);
      vclear = testNode.clear;

      assert_equals(vclear,"none");
       
    domtest.done();
}));

</script>
<p>The clear IDL attribute of the br element must reflect the content attribute of the same name.</p>
<div id="log"></div>
</body></html>
